requirements.txt
google-api-python-client
pytube
moviepy
opencv-python
librosa
mediapipe
speechrecognition
nltk
spacy